requirements.txt
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
Flask==2.2.5
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
google-generativeai==0.3.0
greenlet==3.1.1
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
langchain==0.1.0
langchain-google-genai==0.0.5
MarkupSafe==3.0.2
mongoengine==0.29.1
passlib==1.7.4
PyJWT==2.9.0
pymongo==4.10.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.37
typing-extensions==4.12.2
Werkzeug==2.2.3
wtforms==3.2.1
zipp==3.21.0
